[GN SDK] Add script for paving devices
This script paves a Fuchsia device. If you have a Fuchsia device
connected when you run this script, this will try to reboot your device
into Zedboot and install Fuchsia.

This script and its test are meant to be run from within a project that
is managed by the GN build system and uses the GN SDK. Use generate.py
to create a new project, then run bin/fpave.sh or bin/fpave-test.sh
from that project.

Change-Id: I29fd4db1ebbe55c9038ad91a0dd4c9cfc68cbb20