[a11y] Move view ref ownership to ViewManager.
Moves the view ref ownership out of the semantics tree service so we can avoid multiple view ref copies in the future. This is a pure refactor and no behavior change is expected. All use-cases are covered by existing tests. A followup CL will create a wrapper class for each view ref and its associated lifecycle logic.

The test to ensure that the viewref signal is received by the semantic tree service is no longer necessary as that class no longer deals with view refs.

BUG: 45755
Test: fx run-test a11y_lib_tests, a11y_tests

Change-Id: I887405a17d92e3290102cdab3c787a6cd7ebb67f