[usb-virtual-bus-launcher] Remove unnecessary directory watch
All child directories of /dev/class are always available by the time
devfs is accessible.  Because of the odd properties of /dev/class with
regard to hiding empty subdirectories, it is difficult to maintain the
exact behavior in the current devfs implementation as we migrate to
using libfs.  One behavior that will change is watching on /dev/class,
so just don't do it.  This is the only code that seems to be doing so.

Change-Id: I673ebe893f22a12f30bcd5b3816e60120b5ea377