[GPT] Do not add the dummy device as invisible.
This change modifies the device to not be added as invisible.
Either device_make_visible should be called on the dummy device or the
device should not be added as invisible. This is the contract with the
ddk and gpt driver is violating it. In the case where the device does
not call device_make_visible at all, Device::Controller::Rebind seems
to do the right thing by hanging unless we have introduce timeouts.
A part of the problem with MakeVisible will disappear once all the
drivers are moved to init hooks. However at the moment if this driver
is converted to use init hooks, it will crash, because the current behavior
of not calling device_make_visible is equivalent to not completing the init hook.

Bug:45638

Test:
while [  $COUNTER -lt 25 ]; do
   echo The counter is $COUNTER
   let COUNTER=COUNTER+1
   fx wipe-partition-tables /dev/class/block/002
   fx init-partition-tables /dev/class/block/002
   sleep 10

Change-Id: I83ce4beb02ba5c53dd166d76dacadddbaf575d4e