[debugger] Lookup options for ambiguous inline chains.
Adds an option to symbol resolution to resolve to the beginning or the
end of ambiguous inline chains. This happens if the address is at the
beginning of an inline function. Some callers will want the most
specific inline, and some will want the least (there is guidance in the
options struct about which to use).

Use this option in the disassembler for call instructions and for
the "sym-near" command. In these cases, if the destination address
points to a function that begins with an inline, the result will be the
name of the inline which isn't correct because the call is calling the
outermost physical function.

Clarify some line table comments and add a paranoid check for a 0 file
index (I haven't seen this but it's theoretically possible).

Fix the index_largetest.cc. It is not compiled by default so had
bitrotted.

Change-Id: I78faf494a35b8bfcf311d50770305150b16d5259