[cprng] Fix blocking on missing entropy.
This CL fixes TWO bugs in the zircon cprng. The most important one was
that the cprng would not block even if not enough entropy had been added
to it. The second one was that the accumulated_ variable was not updated
if the cprng was not thread safe, meaning that the entropy added during
early boot would not be counted towards the minimum threshold for
unblocking.

One of the bugs was caused by an inverted condition in the
initialization of the blocking signaling mechanism, and the other one
was due to an early exit from an if condition (the accumulated_ value
was being added to inside the if guard).

This CL adds two tests for this behavior. One of the tests verifies that
the cprng blocks if it doesn't have enough entropy, then adds entropy
and makes sure that the thread unblocks. The other test checks that
there is no blocking if enough entropy has already been added before
becoming thread-safe. These two tests cover the bugs in this CL.

We found this bug with Cameron (eep@) while debugging kcov support for
fuchsia. We got suspicious when we noticed that the coverage was
reported for code paths that waited for the cprng to be initialized, way
after the system had booted correctly.

BUG=45733

Change-Id: I8d04406e09b7f840455d5c12f12fc8ce1f76cb15