[debugger] Fix race when resuming a breakpoint.
Because of the order of the operations, the debugger would resume an
exception before attempting to reinstall the breakpoint, thus
introducing a short window where the thread could execute some
instructions before the breakpoint is re-installed.

In the case of a SW breakpoint, it's rare to hit this as a thread
doesn't normally re-run the same instruction quickly, but in the case of
a R/W watchpoint there are many patterns where two "triggerable"
instructions can be very close: `some_int_variable++` roughly expands to
a read instruction followed by a write instruction.

The fix is to keep the exception handle until *after* the breakpoint has
been reinstalled.

TEST=The verification was made by instrumenting the kernel to get a list
     of the operations done by it. With this we could verify that now
     the installation is indeed done within the exception handle
     lifetime.

Change-Id: I6b59fa184e8696db52899b2adcc89ff37f2bcb4c