[kernel] Fix race between read/write thread state and context switch
This change restructures how and where user-visible register state is
saved in order to eliminate races between debugger and debuggee
threads.

For example, prior to this change arm64's TPIDR_EL0 was stored in a
context switch frame pushed on the thread's stack.  However, there was
no synchronization to ensure that the debuggee had pushed the
frame *before* the debugger attempted to access it.  The result was
that a debugger thread might read a garbage value or, in the case of
trying to write TPIDR_EL0, might corrupt the thread's stack.

For background and more detail, see fxb/33942 and the design document
linked therein.

* Add low-level, arch specific routines for saving/restoring the
  various kinds of user-visible thread state (tpidr, fs_base/gs_base,
  vector, fp, debug regs).

* Add thread_ wrapper functions that call the low-level arch specific
  routines for saving/restoring user-visible state.  These wrappers
  are used to save/restore during thread suspension and exceptions.
  Add RAII helper for installing/removing exception context.

* Change arch_context_switch to skip saving user-visible thread state
  when oldthread is suspended or in an exception.

* Save user state before thread suspension and before sending
  exception messages.

* (arm64) Move storage of TPIDR_EL0 and TPIDRRO_EL0 out of context
  switch frame and into arch_thread_t.

* (arm64) Remove arm64_set_debug_state_for_cpu and defer
  enabling/disabling hardware debug support to kernel entry/exit
  rather than perform it during context switch.

Testing - Re-enabled existing tests and ran core-tests in a loop for
48+ hours on Cavium (qemu+kvm) and on NUC (Dawson Canyon).

Fixed: 33942
Change-Id: Ida8287807d88b95f6393cf9cb6aa9addd00d076c