[graphics][compute][spinel] Unseal styling in gtest fixture
No change in behavior / not a bug.

The gtest fixture only renders the SVG document *once* so it wasn't
necessary to unseal the styling container in order to reuse it in
the next frame.

However, it should be unsealed and reset for clarity.

All tests reliant on this gtest fixture succeed.

Testability:

Target:

 * fx shell run graphics_compute_vk_unittests --gtest_filter=*svg*

Host:

 * out/default/host_x64/exe.unstripped/graphics_compute_vk_unittests --gtest_filter=*svg*

Change-Id: I97ceabea3f8bc7af6b6c98090b01fc6b9842e54f