[gn sdk]Moving python unit test into test project
Previously the python unit test was loading the script under test
from the source tree. A side effect of this was to create a .pyc file
in the source tree, which was sometimes used by another test as a
valid input file. This file is unexpected, and caused the
test_generate.py test to fail.

The test is now moved to the test_project and references the script
under test relative to the test project, which is in a temp dir vs.
the source dir, avoiding this issue.

Change-Id: I42dac483a3327812e63aca30f8be908c6d64a400