[bt] Remove bt-transport-usb device on errors
This causes the bt-transport-usb device to attempt to gracefully
shutdown should the underlying usb transport produce an error.
Although shutting down is not an ideal path, there is no generally
correct way to recover and this is better than attempting to
continue to use a broken usb endpoint.

Without this change the bt-transport-usb would, on the NUC,
occasionally have a thread spinning at 100% cpu. This thread would
be the usb-request-callback-thread and was endlessly re-enquing
requests in hci_acl_read_complete->queue_acl_read_requests_locked,
which would immediately complete due to endpoint being in an error
state.

Test: Booted a NUC many times, observed no threads using 100% cpu
Change-Id: I353041f055610f43395d89783318aa2015928f62