[escher] Add support for pre-warming render-pass/pipeline caches.
Step 6 in a series of CLs to build Vulkan pipelines at startup, rather
than lazily as needed.  Adds:
- per-Frame and per-CommandBuffer disabling of lazy pipeline creation
- utilities to warm pipeline cache per-PaperRendererConfig
- WaterfallDemo now generates all pipelines before the first frame

This functionality is not yet enabled in Scenic; this will be done in
a subsequent CL.  However, this CL enables it in the Escher Waterfall
demo.

Testing: Waterfall demo tests would crash if pipelines weren't properly
pre-generated (verified).

Bug: 44274 #fixed
Bug: 44277 #fixed

Change-Id: Id6a3b8cefd639637c713b25a6a9a6425fc6f4358