[gn sdk] generate vulkan loadable module
Adds handler for loadable_module sdk atoms, but we only support the
"vulkan_layers" atom.

Generates a set of GN targets per layer found in the vulkan_layers
meta.json file. Layers are composed of a json resource file paired
with a binary of the same filename.

Testing: updated testdata and golden files
Change-Id: I65321ff822146c138588248fb780b45ab216f4d2