[audio] Remove the NotifyPosition() domain_token() requirement
We advertise that NotifyPosition() can be called from any thread
however we impose a restriction in the form of requiring the
domain_token(). We can remove this restriction by protecting the
rb_channel_ usage with channel_lock_ (a subsequent CL removes the
rb_channel_ altogether with a completer that is itself locked).

A newer Clang version thread analysis flagged
TaskClosureMethod::CallHandler() as not having the domain_token() when
calling driver's ProcessRingNotification() that required the
domain_token(). TA_GUARDing the timer associated with
the ProcessRingNotification method like this:
async::TaskClosureMethod<MockSimpleAudio, &MockSimpleAudio::ProcessRingNotification> notify_timer_ TA_GUARDED(domain_token()){this};
did not convey the domain_token() guarding for the method to thread
analysis, it was conveying only the guaring of notify_timer_.
All this is mute with this change since we are removing the
domain_token() usage for NotifyPosition().

Fix: 45568

Test: runtests -t device-enumeration-test on Sherlock.
Test: runtests -t sa-unittest-test.
Change-Id: I67af0a2f4179b12847382c7f9333fdfb9cea043e