[components] Hub no longer needs Model nor Realm
This CL removes the last access to Model or Realm from Hub.

on_before_start_instance_async was trying to grab the component_url
from the realm. Instead of forcing the Hub to hold onto a Model to look
up a realm, we just pass in the component_url as part of the
BeforeStartInstance event.

Change-Id: I7307f39658192d0717a49d600b5124187c9827be