[components] Improve error messages on failure of "expect_next".
If a call to "expect_next" requests more elements than are present,
we we currently panic in the "drain" call.

Instead, only request at most "max(desired, actual)" elements. If we
didn't receive as many events as we expected, we will fail the
"assert_eq" with a more helpful error message.

Bug: 45296
Change-Id: I5c2417da3e699857503eb22bc81423baf725c2f1