[gn sdk]Propagating metadata for all atoms
This forwards the metadata for each of the atoms included in
the GN SDK. This allows the GN SDK to be processed in the same
manner as the Core SDK; by processing the meta/manifest.json.

The additional files included in the GN SDK are added to the
manifest.json so that no files are included in the SDK that
are not enumerated in the manifest.

Test is added to verify that all files in the SDK are referenced
in the manifest.

Change-Id: I41afe7f5c076bfe018ca763b6b08f089c3c5f705