[i18n] Adds icu configuration data for Chromium
This caused the tests for Chromium not to be ran where the test names
were `base_unittests`, which meant the test would be silently skipped
on local runs.

We need separate configurations for semantically two identical tests.
This is because configurations are provided based on package name, and
the chromium tests get ran under different names in different
circumstances.  If both configs end up included in a product, they will
not take much extra space, as they are hash-identical and will be
deduped at rest.

Bug: 45491, chromium:1047475
Change-Id: I98398152cfd21ca1ef79977665eb3a7543d851f1