[gpt] 0-init mbr_partition_t in FinaliseAndSync
mbr_partition_t was not being 0-initialized, and therefore the `status`
field was not being set correctly. This could cause the protective MBR
around a GPT to be marked as active, and cause an EFI bootloader to
ignore the ESP on the disk.

To recover a device, make sure you run `gpt init /dev/...`.

Bug: 45625
Change-Id: Ia14cf271aa0fb66003afebe4dfb81a008ac6d28d