[kernel][crashlogs] Fix kernel panic crashlogs.
During a recent refactor, kernel panic crash-logging got a bit broken.
This was caused by a move of dlog_shutdown down from the graceful
platform halt level to the new common platform halt level.  This is
not a safe function to call when interrupts are off (as they are
during a panic).  The primary reason for this movement of
dlog_shutdown was to make sure that the logs were flushed to make a
QEMU host test (the oom_test) happy.  For now, instead of making the
test happy by attempting to log the string it is looking for at a low
level, log it at the high level in the OOM handler just before it
calls the platform halt helper.

Tests: On Sherlock
> k crash
> k pmm oom
> k wdt force

I'm relying on CQ to run the oom_test.  No matter how hard I seem to
try, I'm not able to get my workstation to run it (or even acknowledge
that it exists)

Change-Id: I5a0b372e96b5250107f7ed661a58024c4a82125c