[debugger] Switch back to Android unwinder.
ngunwind gets lots on "operator new" in optimized ARM code which comes
up a lot. The result is that stepping frequently gets confused.

Previously we switched back to ngunwind due to problems in syscalls. The
main problem was with assertions but that was due to a compiler issue
that has been worked around in most cases.

Change-Id: I64dbec5a4cbd5ba548e5bae7064846b56eadfe9f