[Brightness] Fix flaky brightness table tests
These tests would flake as they can be run simultaneously.
An easy way to reproduce would be to add a multi-second thread::sleep()
call to the start of test_read_brightness_table_file_empty_file(),
forcing it to be run last and consistently causing it to fail.
Change each test to use its own file to avoid the race condition
and flakes.

Bug: 45873
Change-Id: If9f6d7d510e81c50726e9ae464307e0bf829a4df