[session] Added capabilities for cast_agent
This adds a few additional services needed by cast_agent that were not
already included in fxr/358314.

Sorts protocols and directories to make it easier to compare and merge
future changes.

Slight formatting cleanup for consistency and ensure trailing commas.

Adjusted comment positioning to make sure the comments stay associated
with the thing they are commenting. Moved comments in lists to the end
of the line so they remain attached when sorting. Moved object (in
braces) comments inside braces so they move with the objects when
reordering.

Test: N/A

Change-Id: I6f12cd447255ab0de26df372e330185e6fc7976c