[appmgr][component_events] Fix issue and add integration tests with observer
Bug cause by leftovers of previous assumptions in which we thought about special
casing root, replying with realm paths that include the realm (not relative) together with
missing tests that properly exercised this behavior.

Bug not marked as fix as we would still like to do an integration test
for lifecycle events probably.

Tested: fx run-test component_events_integration_tests
Bug: 43183

Change-Id: I221d1812f23a4b627d2d91ad75daa7c04accfd81