[files_async] Allow multiple calls to readdir and readdir_recursive
This change ensures that the directory offset for
a directory proxy is reset before attempting to
read the directory entries.  This fixes behavior
where multiple calls to readdir results in the
correct list of entries the first call, and an
empty list in subsequent calls.

Testing: fx run-test files_async_tests

Change-Id: Id49a63de3b14845018bd47901fc7f0b3a68990d2