[debugger] Fix arm problems.
Clear compilation directories when looking up line information that
results in "0" line entries. The previous behavior resulted in a
FileLine with no line number and no file, but with a compilation
directory which makes no sense.

This is enforced with a new assertion since we should never have this.

Use the new "outer" option for resolving disassembly call destinations.
In cases where the destination is an ambiguous address, usually the user
will want to see the outermost function name.

Change-Id: I02ac52137f7c39bb5f379a07555cd4cd52bca0a1