[wlan-hw-sim] Require primary channel for beacons
The existing API does not prevent anyone from sending a beacon on every
single channel, which can be confusing. Given the fact that one BSSID
should only be broadcasting beacons on one channel, it makes sense to
require a primary channel to be specified for any given fake beacon and
only send beacon on that channel.

Bug: 44899
Test: Test-only change. Does not change behavior

fx run-test wlan-hw-sim-test

Change-Id: Ia4ce61aa12925996fc36780df1868b82893cd89a