[sktool] Add transaction handling.
This adds the first complete transaction (i.e. a send of one or more packets
followed by a receive of one of more packets) implementation, complete
with some connection error handling if things go off the rails.

It also moves the implementation of ping over to the transaction
implementation instead of asking it to craft its own packets directly,
and since we may need to generate random bytes from our RNG during a
recovery we need to add mutable internal state for the first time.

Bug: 44796

Change-Id: I89f9ec30432e87aa9e69344519fcc9748045f852