[bindc] Add spans to the Statement and Condition AST nodes.
This CL adds a span property to the Statement and Condition AST nodes.
These nodes were chosen since their spans will be needed by the offline
debugger, but it is likely that spans will eventually be added to other
nodes as well. This involved changing the input type of all parser
functions to a LocatedSpan struct provided by the nom_locate library.

Change-Id: I77caac2634d12ebc566d1c5cfdb6af2c3d41f640