[audio][virtual_audio] Change some domain token's usage.
A newer Clang version thread analysis flagged
TaskClosureMethod::CallHandler() as not having the domain_token() when
calls are made to various virtual_audio TaskClosureMethods.
The methods are safe since the Posts are made to the one dispatcher().
For these methods remove the domain_token() TA requirement and added
domain acquisition with comments per the description above.

Fix: 45568

Test: Build, compile time locking spec change with no behavior change.
Change-Id: I632ea096fe1f5423db9f5eb8fc9fc7b0565ea912