[component_manager] add dependency_type field to offers
Adds a new field called "dependency_type" to protocol and directory
offers. This will be used to mark dependencies which should be ignored
by the shutdown process, thus breaking dependency cycles.

This change adds the new syntax to cml/cm/fidl, a followup change will
implement the logic in component manager to use the new field.

Change-Id: I54203b033e9d48bffc23c924ab223fc27a9bd4fb