[appmgr] Removes data race
There is a data race between the destructor of
ComponentEventProviderImpl and NotifyOfExistingCompoments.  The latter
call could come in when `this` is invalidated, leading to an asan crash,
and therefore to test flakiness in an asan build.

Since the code is single-threaded, a weak pointer to this class is
enough to ensure callback is never invoked on an invalid pointer.

Tested:
  fx run-test appmgr_integration_tests

Bug: 45989
Change-Id: I2e9a8aed9fd2cf2f958f39a91db3cb2230d9b018