[system_updater] Parse system update packages.json
This change modifies the system updater to read a packages.json file
when populating packages. If there is no packages.json file present,
we'll fall back to reading the original line formatted packages file.
This is important because the json file allows us to specify
fully-qualified URIs.

Note: metrics will be added in a later CL.

Bug: 44774
Change-Id: I9bc4ebf417ab6d27532a1a380eb8186038d8fe6e