[syscalls] Add zx_system_get_version_string to replace zx_system_get_version
Establish a new "primitive" type for vDSO functions that return
read-only strings embedded in the vDSO itself (zx_string_view_t).
This type can't be represented meaningfully in FIDL but is a perfect
fit for many languages' native immutable string view/slice types and
is optimally cheap to return as a function return value in the C ABI.

Use this for the version string syscall, which gets a new name.  The
old syscall that fills in a buffer is now deprecated and will be
removed when all users have been migrated.  It's not annotated for
compile-time deprecation warnings because out-of-tree users do not
tolerate the warnings.

Also add an idiomatic Rust binding for the new syscall.

Fixed: 45640
Change-Id: Ic20efc451ce4f3b2f742c5eec88866425166ba4e