[camera] Add fake camera lib
This change adds a fake for the fuchsia.camera3.Device interface. Users
can create a device using fake streams, and register callbacks to
participate in buffer negotiation from the perspective of the device.
This can be used in conjunction with the fake stream interface to create
virtual cameras with software-defined stream content.

Test: added FakeCameraTest suite
Change-Id: Ib501920acd36c4158721fa49f89c0b721be6b279