[i18n] Implements stand-alone fuchsia.intl.
This code repurposes the modular implementation of `fuchsia.settings.*`
into a stand-alone binary, `intl_services.cmx`.

Since I also discovered that `fuchsia.setui.SetUiServices` is actually not
supposed to even work with anything that are not account settings, used
this opportunity to migrate from `fuchsia.setui` to `fuchsia.settings`.
This looks like it will remove a host of bizarre issues observed around
the use of `fuchsia.setui`.

In the process, however, I noticed that the `fuchsia.settings.Intl` does
not have full feature parity with `fuchsia.setui` as it does not support
the hour cycle field.  Added this in a separate commit. An implementation
of hour cycle support in `setui_service.cmx` is pending.

Tested:
  fx run-test modular_integration_tests
  fx run-test intl_services_test
  fx run-test intl_property_provider_impl_test

Bug: 45216
Change-Id: I860755ece007adaf00fe5d0e7f82ebbdcd384b43