[scenic][flatland] Gather instance data in an "UberStruct".
In the future, each Flatland instance instance will run on its own
thread. The render loop (also on it's own thread) will process a
snapshot of the data most recently presented by each Flatland
instance to render the next frame - this CL makes that data live in
the "UberStruct".

For now, the UberStruct only contains a local topology for each
Flatland, but in the future will contain other Flatland-local data
such as matrices and references to resources like images.

The UberStructSystem provides lock-protected setters for each
instance to modify its associated UberStruct, as well as a function
to snapshot the current state of all UberStructs. The intended flow
is as follows:
- Flatland instances set a new UberStruct at the end of every
  successful Present() call.
- The render loop snapshot the entire state of the UberStructs at
  the beginning of each frame to ensure nothing changes mid-render.

The TopologySystem, which previously held onto the data that now
lives in UberStructs, has been broken into two parts: the topology
storage is the UberStructSystem, and global topology compuations were
moved to the TransformGraph.

Additionally, this CL reworks how link topologies are represented.
Rather than doing local topology expansion for any Transform, only
"link transforms", as identified with a unique link instance ID, will
be checked for expansion. The LinkSystem stores the active list of
topology links and provides a snapshot as another piece of data the
render loop will consume.

TEST=fx run-test scenic_tests -t flatland_unittests
Bug: 44334
Fixed: 42583

MULTIPLY: `[
  {
    "name": "scenic_tests",
    "os": "fuchsia",
    "total_runs": 30
  }
]`

Change-Id: Id5c11947543c2d91c5ce56b1fb429c8e4f82e9c2