[cmc] Produce an error in CML files if no runner specified.
CML files that have a "program" block but don't "use" a runner are
likely incorrect: the user presumably expects it to run the defined
binary, but --- because there is no runner --- the component will
never actually have a program run.

Instead of silently treating such files as a component without
a program, generate an error at compile time. CML files without
a "program" block (and hence, should have no user expectation of running
any code) continue to be allowed.

Bug: 45890
Change-Id: I7a4d313cb3fd8bbcbd1923bed42d41fb058624aa