[paver] Add support for AVB to EFI partitioner
Add support for the VBMeta and ABRMeta partitions to the EFI partitioner
in order to support AVB.

Additionally, this change also make EFI and CrOS partitioners work like
the Sherlock partitioner, and wipe known partitions during GPT
initialisation.

Finally, this change also doubles the default size of the FVM partition,
so we are larger than the minimum MinFS size required by the CrOS disk
setup code.

Test: Ran fx pave on NUC and Pixelbook Go, both before and after
      re-initialising the partition tables. Expanded tests and added new
      tests (but kept them all disabled).
Bug: 46098
Change-Id: I2aa99184ef6607e16a4c6b8cd3a21eca0eed7705