[disk-inspect] Add basic cl string command parsing library.
Minimal prototype library to handle parsing a command arg string
from the command line to a handler to perform inspection operations.
Disk inspector needs to be able to multiplex many different custom
set of commands unique to each filesystem. This library provides a
Command object for the input reading application to wrap a command
in and a CommandHandler interface for fs libraries to implement to
expose what commands are supported and to handle passed in commands.

Bug: 37907
Change-Id: Ibb6ffe228feafec0d82b867a60f4bc873a9581f1