[ramdisk] Complete all transactions on detach
I can not figure out how to write a proper test for this.
Steps proposed by fxb/38162 expose the bug, but they are too heavy
weight to use as a test for ramdisk, as they involve blobfs code.

Seems like a previous change in this area:

    commit 1019f5e868d08624b1f571878e467d8cbc88740d
    Author: Suraj Malhotra <surajmalhotra@google.com>
    Date:   Sun Jul 21 16:57:31 2019 -0700

        [ramdisk] Fix flake due to not completing

        This was missed when changing operation lib behavior to no longer
        autocomplete in destructor and resulted in a flake.

        Tested: runtests -t ramdisk-test x100
        Change-Id: I055f9dea07427fd6c9e5143448d145bd14dacea0

does not include a unit test either.

Tested:
  - apply fxr/326996
  - run: fx set core.x64 --with=//src/sys/pkg/testing/blobfs-ramdisk:blobfs-ramdisk-tests
  - run: fx run-test blobfs-ramdisk-tests -- trigger_ramctl_hang --nocapture

Fixed: 38162

Change-Id: Ie8cfe472cc2de8e7fc7b07e42b62b66bdc419109