[graphics][compute][spinel] Lost blocks in ring buffer
Bug:46116

A work-in-progress path might be skipped if the current work-in-progress
(unlaunched) dispatch is forcibly launched due to either running out of
resources (e.g. acquiring a ring block) or an explicit flush.

This will result in a lost path and downstream corruption.

The root cause is that after an implicit or explicit flush, a new
"dispatch" record is acquired which instead of pointing to the first
block of a work-in-progress path, it was pointing to the next block in
the ring.

This bug would only reveal itself in two cases:

 1. Invoking `spn_path_builder_flush()` during construction of a path
 2. Rendering a very very large path with many dispatches in flight

The unit test recreates case (1).

Patch set 1 will assert and fail.

Patch set 2+ fixes the issue with an improved path builder.

A debug-only assert() that ensures "conservation" of the ring's blocks
is left in place.

This CL also judiciously updates comments in path_builder_impl.c.

Testability:

Test is `spinel_vk.wip_path_is_lost`

Target: fx shell run graphics_compute_vk_unittests --gtest_filter=*spinel*

Host: out/default/host_x64/exe.unstripped/graphics_compute_vk_unittests --gtest_filter=*spinel*

Change-Id: Ia39a30d9fe83941ffa4bbc238bdaeb4e6f4c267a