[audio_core] Slab allocate Packets
This are allocated and free'd very frequently. Migrate to a slab
allocator to prevent heap fragmentation. Additionally the upper-bound on
the number of instances will be a signal if we ever have a memory leak
with these packets.

Each renderer now has its own slab allocator which sets an upper bound
on the number of outstanding packets the client may send at any given
time. The slab allocator is tuned to allow 584 packets per client, or
64kb. If we are unable to allocate a new Packet from the renderers slab
allocator, then the renderer will be disconnected.

Bug: 35709
Test: fx run-test audio_pipeline_tests
Change-Id: I50ceba4108e46b1a7fde66e1e447dc3dde67fc95