[syscalls][serial] Support partial disabling of serial syscalls
Extend kernel.enable-serial-syscalls to also support an "output-only"
setting, in addition to the previous fully-enabled and fully-disabled
values.

This supports enabling serial write (from the target's perspective), but
disabling read.

Extending the previous command line value allows for a soft transition
for configurations where =true and =false are already in use.

Test: new test in disable_syscalls_test
Bug: 44751
Change-Id: Ie3f937ee1e8d562258ee73dc442a44dd2039bfc6