[audio][astro][sherlock] Abort driver init upon codec failure
Astro and Sherlock drivers did not check for failures returned
by codecs at initialization. This was uncovered by a screeching
sounds heard from Sherlocks when the I2C bus driver used to
communicate with the codec went down due to an unrelated failure.
This CL adds checks for return values that boil down to status
from I2C calls that would have prevented the audio drivers from
completing initialization such that the audio subsystem was never
enabled for clocking data in the TDM bus when the codec is not
correctly initialized.

Also for Sherlock, the codec had a check for kRegFaultCfgErrorStatus
(register 0x08) which would alway error out because the codec is
configured before there is correct clocking, so this check was
removed from codec initialization.

Fix: 45518

Test: runtest -t astro-audio-out-test
Test: runtest -t sherlock-audio-out-test
Test: runtest -t device-enumeration-test on Astro and Sherlock
Test: audio uapp tone on Astro and Sherlock
Test: Fake I2C failure by crasing the process were I2C bus lives and
check that the audio driver is not loaded via dm dump, this on Sherlock.

Change-Id: I0d6c7349ac711d17d719fc3931c2ed12284fac8e