[blobfs] Fix flake in blobfs checker unit test.
Running TestAllocatedInodeCountTooHight and
TestAllocatedBlockCountTooHigh on an empty blobfs is flaky. Running
instead on blobfs with at least one blob in it seems to
resolve the flake and from checker's point of view it tests the same
code path.

This change also uses a helper for writing the superblock. The prior
implementation failed to call DetachVmo and the helper addresses that
oversight.

BUG=45770

Change-Id: I521b6564920a12803939b4a6c715005cd2b47447