[sync][completion] Add a state check back in to avoid spurious wakeup.
Put back in a state check which was removed as part of SHA
51585402e831516319adf13276b9f2e984a065c9.  Basically this undoes the
trade made in that CL allowing for hypothetical "missed" events, but
avoiding very real spurious wakeups.  See bug 45420 for details.

Also, undo the workaround which was put in place previously.  This did
not actually do anything aside from slightly changing timing to
address the problem.

Tests:  Basic qualification, and one 3 day weekend of focused stress
testing with the test case used to investigate the problem.  This
involved 8 x64 KVM instances and 32 ARM64 KVM instances.  Also added a
weak unit test.  It is not guaranteed to find problems, but if it
finds one, the problem should be real (eg. no flakes.)

Bug: 45420
Change-Id: I13d0c7b6d9ec0aabfa4683ba2e15077759b78e9f