[audio_core] add thermal agent
This CL adds, but does not integrate, a thermal agent into
audio_core. The thermal agent is configured using the pipeline
config. If the pipeline config contains any entries specifying
response to thermal changes, the agent registers with the thermal
controller and issues (using a callback) config changes.

Later, the thermal agent will be integrated with the rest of
audio_core so that the config changes have real impact on the
effect instances in the pipeline.

TEST: unit test added for the thermal agent

Change-Id: I89af4402285d3c8bc27add66fb39a5c4102f7f21