[components] Add `core` component under `root`
`core` is the rough CFv2 equivalent to appmgr's `sys` realm. It
provisions capabilities than are used generally through the rest of the
system. `core` is a non-executable component that lives in its own
package.

`sys` environment components from appmgr will usually be migrated to
live under `core`. `appmgr` itself also lives under `core`. This
topology makes it straightforward to arrange a migrated system component
in the CFv2 topology: just make it a child of `core`, and route services
to/from other components in `core`.

Change-Id: Ib759ca69e1ee1efcae40b1e705caa17a3197e19c