[audio_core] Fix possible packet leak
In DiscardAllPackets we were recomputing our PTS to fractional frame
mapping to a 0 PTS value. For clients that use timestamped packets, this
will cause the mapping to be incorrect with respect to the client
packets. This results in audio_core retaining these packets until the
associated PTS value, which can easily be minutes or more depending on
the PTS values the client selected.

This is particularly noticable for clients that timestamp packets using
the absolute system clock value, where packets can be held for hour or
days (depending on uptime) which can result in runaway heap memory
usage (or disconnection of the client after fxr/35709).

This logic to mutate the PTS to frames transform is not needed anyways,
so we simply remove the logic and update our tests to validate the new
behavior.

Bug: 46136
Test: fx run-test audio_pipeline_tests
Change-Id: I5cd9b0faa96565651184f6f6c6def71a97ad5bb3