[dhcpd-cli] Add BoundDevices
Add BoundDevices as a variant of Parameter. This change adds support
to dhcpd-cli for configuring the network devices to which dhcpd binds.
No tests have been added because there is nothing meaningful to test in
this change; a test would merely verify that the argh crate and the rust
type system work as expected.

Change-Id: Ib294ab2a42b888b24241e3906dd03e0341081bf2