[build][fuzzing] Add "fuzzer profiles" support.
This CL adds a `fuzzer_profiles` parameter to the `fuzzers_package`
template. This parameter includes the original parameters of the
template. Its intent is to separate the specification of how to build
sets of fuzzers from how to install sets of fuzzers. In particular, a
`fuzzers_package` can now specify different sanitizers for different
fuzzers.

This is particularly useful as we move to support multi-lingual
fuzzing. For targets that involve two or more toolchains, there are a
correspodning number of ways to instrument the code being fuzzed. For
example, a fuzzer written in a combination of Rust and C/C++ may use
rustc to sanitize the rustc code and link it against its compiler
runtime, or clang to sanitize the C/C++ code and link it against *it's*
compiler runtime.

This CL allows fuzzers using distinct toolchains to be logically
grouped:
  fuzzers_package("some_crate_fuzzers") {
    fuzzer_profiles = [
      {
        fuzzers = [ ":unsafe_cpp_fuzzer" ]
        sanitizers = [ "asan", "ubsan" ]
      },
      {
        fuzzers = [ ":rust_code_fuzzer" ]
        sanitizers = [ "rust-asan" ]
      }
    ]
  }

Bug: 24965
Change-Id: Id388b4cf44bc76d1f6da62ec1a99df29e1555080