[ftl] Convert each part of the RD32_XE macros to ui32 before shifting
UBSan was complaining about left shifts of 255 by 24 not fitting in an
int. Although each shift is operated on a ui8, and unsigned type
overflow is defined, each of the shifted values is actually promoted to
an int before the shift occurs, hence the overflow complaint.

This CL just explicitely converts each part to ui32 to allow the defined
unsigned overflow.

Bug: 41889
Change-Id: I0930928c1e98d4013ca5e81ee678327e3207574a