[build] remove xi
We have been building but not using xi for a long time. Our mirror is
now out of date, and builds occasionally fail which is undesirable.
Removing these builds saves us significant build time, so until we
are once again maintaining Xi for fuchsia, removing this has plenty
of advantage and little disadvantage.

Change-Id: I3b0c97e266fb82a7ab983f103a7243034bb1487a