[i18n] Add ICU config data for web_runner.
Adds the configuration needed so that web_runner can load updated tzdata
files if available.  This will also remove a harmless but distracting
error message in the logs about the lack of these files.

This change alone does not make the config available or the message go
away. Additionally we need to hook up the configuration to the product
definition so that it becomes available to Chromium builders.

Bug: 45462
Change-Id: I361d103b1f365b9dc9fe0e4693f5fcf7859c218b