[debugger] Ignore casting errors for auto upcasting.
When auto upcasting to derived classes, some things can fail. In
particular, virtual inheritance makes this impossible.

Previously this would forward the casting error and the attempt to
auto-upcast would report an error. But we want failures to auto-upcast
to just default to the same value as original (fall back to no
auto-upcasting).

Adds a fix and test for this.

Removes an unused parameter in VirtualInheritanceTestSetup.

Change-Id: I77cb91d2f9950f9aad58d3f1c3da01b969442259