[kernel][stream] Implement append for streams
Originally, this feature was specified as a mode on the stream object,
but trying to use the feature in memfs, it would be much more convenient
as a flag on writev. Given that we already have an |options| parameter
to writev, I've implemented the feature that way and updated the docs.

Also, implement the ZX_INFO_STREAM topic, which is exercised in the test
for appending to streams.

Change-Id: I518102e93da95fb40e66810d5945cef11bbc07ce