[netstack3] Do not validate Ethernet frame length
Previously, we validated that inbound Ethernet frames adhered to
the required minimum length defined in the specification. However,
this caused problems in practice since, at least on Linux, Ethernet
frames are only padded to the minimum length when they are sent
over physical media, but not when sent over virtual devices.

We have decided that the driver or the hardware are more appropriate
places to perform this validation, and that the netstack should
accept smaller frames if they are received by a driver which allows
them.

This change does two things. First, it extends the Ethernet parsing
utilities in `wire::ethernet` to allow the user to decide whether
or not to validate the frame length during parsing. Second, it
updates the code in `device::ethernet` to make use of this to opt
out of frame length validation.

Test: Some test cases added to the tests in `wire::ethernet` to
      check for both the length-validating and non-length-validating
      cases; run with ` fx run-test netstack3_tests -t
      netstack3_core_lib_test`

Change-Id: I58d2c1752fa08822091d2466d27a6d8d7f43d688